[IMP] pms_w_s: use BookingEngineParser in /rooms
